requirements.txt
plotly
Pillow
dash
dash_bootstrap_components
dash_daq
openpyxl
xlrd
